fix: 폴더 생성시 input이 사라지는 문제 수정
