+++ frontend/techpick/src/features/nodeManagement/hooks/useTreeHandlers.ts
@@ -161,40 +161,6 @@ export const useTreeHandlers = () => {
     await refetchStructure();
   };
 
-  const handleRestore = async ({
-    ids,
-    nodes,
-  }: {
-    ids: string[];
-    nodes: NodeApi[];
-  }) => {
-    const realNodeId =
-      nodes[0].data.type === 'folder'
-        ? nodes[0].data.folderId
-        : nodes[0].data.pickId;
-
-    const updatedRoot = structuredClone(structureData!.root);
-    updatedRoot.splice(0, 0, nodes[0].data);
-
-    const updatedRecycleBin = removeByIdFromStructure(
-      structuredClone(structureData!.recycleBin),
-      ids[0]
-    );
-
-    const serverData = {
-      parentFolderId: defaultFolderIdData!.ROOT,
-      structure: {
-        root: updatedRoot,
-        recycleBin: updatedRecycleBin,
-      },
-    };
-    await moveFolder({
-      folderId: realNodeId.toString(),
-      structure: serverData,
-    });
-    await refetchStructure();
-  };
-
   const handleDelete = async ({
     ids,
     nodes,
@@ -232,6 +198,5 @@ export const useTreeHandlers = () => {
     handleRename,
     handleMoveToTrash,
     handleDelete,
-    handleRestore,
   };
 };
